Database changes!
            //Convert Event IsGlobal to bool and rename to Global - Done?
            //Change EventPage Graphic Type to an enum instead of an int - Done?
            //Change EventPage Layer, MovementType, MovementFreq, MovementSpeed, Trigger to enum
            //Make AddChatboxText Channel an Enum
            //Make Change Gender Command Gender an enum
            //Make Set Access Command Access an enum
            //Make player/server variables longs
            //Maybe make Warp Dir an enum
            //Remove TargetId from EventMoveRouteCommand
            //Rename Spell CastingReqs to CastingRequirements
            //Change Quest LogBefore/After/Quitable/Repeatable to bools
            //Rename Options MapWidth/Mapheight to just width and height
            //Disable api by default -- set api port to server port by default
            //Change map npc spawn dir to byte
            //Rename any Desc to Description


Former-commit-id: 0b9b5db6166220366c5b074a6c3f786b5ee8fb38

+++ Intersect Server/Classes/Entities/Entity.cs
@@ -1689,7 +1689,7 @@ namespace Intersect.Server.Classes.Entities
             }
             if (GetType() == typeof(Player)) //helps the client identify admins if entity is a player.
             {
-                bf.WriteInteger(((Player)this).MyClient.Access);
+                bf.WriteInteger((int)((Player)this).MyClient.Access);
             }
             else if (GetType() == typeof(Npc)) //Helps the client identify NPC Behavior
             {
